added app icons and new logo. changed balance value to float.

+++ LabsMobile/LabsMobile SDK/LMOBalanceResponse.h
@@ -23,6 +23,6 @@
 /**
  The credit of the account.
  */
-@property NSInteger messages;
+@property float messages;
 
 @end
